LPS-196082 Inline GroupPermissionImpl logic into Util
